My_AI_Employee/pyproject.toml
[project]
name = "my-ai-employee"
version = "0.3.0"
description = "Gold Tier AI Employee - Autonomous operation with business intelligence"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Bronze Tier dependencies
    "pytest>=9.0.2",
    "python-frontmatter>=1.1.0",
    "python-dotenv>=1.0.0",
    "watchdog>=6.0.0",
    # Silver Tier dependencies
    "fastmcp>=0.1.0",
    "playwright>=1.40.0",
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.1",
    # Gold Tier dependencies
    "odoorpc>=0.9.0",
    "facebook-sdk>=3.1.0",
    "tweepy>=4.14.0",
    "psutil>=5.9.0",
    # Required utilities
    "keyring>=24.3.0",        # Secure credential storage (used in utils/credentials.py)
    "pyyaml>=6.0.1",          # YAML parsing (required by python-frontmatter)
    "python-dateutil>=2.8.2", # Better date parsing for CEO briefing
    "schedule>=1.2.0",        # Job scheduling for weekly CEO briefing
    "requests>=2.31.0",       # HTTP library for token regeneration and API calls
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["watchers*", "utils*", "vault_ops*", "triage*"]
exclude = ["AI_Employee_Vault*", "test_watch_folder*", "tests*"]
